add StringValidatorAttribute which allows task developers to specify whether an empty string value is allowed for a certain task attribute.
Currently, when the Required property of the TaskAttribute(Attribute) is set to "true", no check is being performed whether the attribute on the build element contains a value or not, which means specifying an empty value for an required attribute on a task would allow it to pass the Required check.

We have considered changing the behaviour of the Required attribute so that the test would also fail if an empty value was specified, but this could break the behaviour of existing (third party) tasks, so we decided against this.

Instead, we've added a StringValidator(Attribute) that has an AllowEmpty property (which is true by default) that can be set to "false", in order to have the build fail when an empty value is specified for an attribute that requires a (non-empty) value.